(Qt) Merge ui_qt_load_core_window.cpp into ui_qt.cpp - further binary size reduction
